+++ jopa-impl/src/main/java/cz/cvut/kbss/jopa/query/mapper/FieldResultMapper.java
@@ -1,16 +1,14 @@
 /**
  * Copyright (C) 2016 Czech Technical University in Prague
- *
- * This program is free software: you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software
- * Foundation, either version 3 of the License, or (at your option) any
- * later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
- * details. You should have received a copy of the GNU General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ * <p>
+ * This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public
+ * License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later
+ * version.
+ * <p>
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied
+ * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+ * details. You should have received a copy of the GNU General Public License along with this program. If not, see
+ * <http://www.gnu.org/licenses/>.
  */
 package cz.cvut.kbss.jopa.query.mapper;
 
@@ -21,14 +19,10 @@ import cz.cvut.kbss.jopa.sessions.UnitOfWork;
 import cz.cvut.kbss.jopa.utils.EntityPropertiesUtils;
 import cz.cvut.kbss.ontodriver.ResultSet;
 import cz.cvut.kbss.ontodriver.exception.OntoDriverException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.util.Optional;
 
 class FieldResultMapper {
-
-    private static final Logger LOG = LoggerFactory.getLogger(FieldResultMapper.class);
 
     private final String variableName;
 
@@ -74,10 +68,12 @@ class FieldResultMapper {
 
     Optional<Object> getVariableValue(ResultSet resultSet) {
         try {
-            return Optional.ofNullable(resultSet.getObject(variableName));
+            if (!resultSet.isBound(variableName)) {
+                return Optional.empty();
+            }
+            return Optional.of(resultSet.getObject(variableName));
         } catch (OntoDriverException e) {
-            LOG.warn(e.getMessage());
-            return Optional.empty();
+            throw new SparqlResultMappingException(e);
         }
     }