+++ jopa-impl/src/main/java/cz/cvut/kbss/jopa/query/mapper/VariableResultMapper.java
@@ -1,16 +1,14 @@
 /**
  * Copyright (C) 2016 Czech Technical University in Prague
- *
- * This program is free software: you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software
- * Foundation, either version 3 of the License, or (at your option) any
- * later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
- * details. You should have received a copy of the GNU General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ * <p>
+ * This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public
+ * License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later
+ * version.
+ * <p>
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied
+ * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+ * details. You should have received a copy of the GNU General Public License along with this program. If not, see
+ * <http://www.gnu.org/licenses/>.
  */
 package cz.cvut.kbss.jopa.query.mapper;
 
@@ -47,12 +45,15 @@ class VariableResultMapper implements SparqlResultMapper {
      * configuration represented by this instance.
      *
      * @param resultSet Query result set to read
-     * @param uow
+     * @param uow       UnitOfWork instance
      * @return The mapped value
      */
     @Override
     public Object map(ResultSet resultSet, UnitOfWorkImpl uow) {
         try {
+            if (!resultSet.isBound(name)) {
+                return null;
+            }
             final Object value = resultSet.getObject(name);
             if (!void.class.equals(targetType)) {
                 return DatatypeTransformer.transform(value, targetType);