+++ jopa-impl/src/test/java/cz/cvut/kbss/jopa/query/mapper/VariableResultMapperTest.java
@@ -1,20 +1,17 @@
 /**
  * Copyright (C) 2016 Czech Technical University in Prague
- *
- * This program is free software: you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software
- * Foundation, either version 3 of the License, or (at your option) any
- * later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
- * details. You should have received a copy of the GNU General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ * <p>
+ * This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public
+ * License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later
+ * version.
+ * <p>
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied
+ * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+ * details. You should have received a copy of the GNU General Public License along with this program. If not, see
+ * <http://www.gnu.org/licenses/>.
  */
 package cz.cvut.kbss.jopa.query.mapper;
 
-import cz.cvut.kbss.jopa.exception.SparqlResultMappingException;
 import cz.cvut.kbss.jopa.model.annotations.SparqlResultSetMapping;
 import cz.cvut.kbss.jopa.model.annotations.VariableResult;
 import cz.cvut.kbss.jopa.sessions.UnitOfWorkImpl;
@@ -29,8 +26,7 @@ import org.mockito.MockitoAnnotations;
 
 import java.net.URI;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.*;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
@@ -55,6 +51,7 @@ public class VariableResultMapperTest {
     @Test
     public void mapReadsValueFromResultSetByVariableNameAndReturnsIt() throws OntoDriverException {
         final Integer value = 117;
+        when(resultSet.isBound(NAME)).thenReturn(true);
         when(resultSet.getObject(NAME)).thenReturn(value);
         final VariableResultMapper mapper = new VariableResultMapper(NoType.getVariableMapping());
         final Object result = mapper.map(resultSet, uowMock);
@@ -75,6 +72,7 @@ public class VariableResultMapperTest {
     @Test
     public void mapCastsValueToTargetTypeWhenPossible() throws OntoDriverException {
         final Integer value = 117;
+        when(resultSet.isBound(NAME)).thenReturn(true);
         when(resultSet.getObject(NAME)).thenReturn(value);
         final VariableResultMapper mapper = new VariableResultMapper(WithTypeCast.getVariableMapping());
         final Object result = mapper.map(resultSet, uowMock);
@@ -95,6 +93,7 @@ public class VariableResultMapperTest {
     @Test
     public void mapTransformsValueToTargetType() throws OntoDriverException {
         final String value = "http://onto.fel.cvut.cz";
+        when(resultSet.isBound(NAME)).thenReturn(true);
         when(resultSet.getObject(NAME)).thenReturn(value);
         final VariableResultMapper mapper = new VariableResultMapper(WithTypeTransform.getVariableMapping());
         final Object result = mapper.map(resultSet, uowMock);
@@ -113,12 +112,11 @@ public class VariableResultMapperTest {
     }
 
     @Test
-    public void mapThrowsMappingExceptionWhenVariableIsNotFoundInResult() throws OntoDriverException {
-        final String message = "No result binding found for label" + NAME;
-        when(resultSet.getObject(NAME)).thenThrow(new OntoDriverException(message));
-        thrown.expect(SparqlResultMappingException.class);
-        thrown.expectMessage(message);
+    public void mapReturnsNullWhenVariableIsNotBoundInResultSetButIsNotRequired() throws Exception {
+        when(resultSet.isBound(NAME)).thenReturn(false);
+        when(resultSet.getObject(NAME)).thenThrow(new OntoDriverException("Error"));
         final VariableResultMapper mapper = new VariableResultMapper(WithTypeTransform.getVariableMapping());
-        mapper.map(resultSet, uowMock);
+        final Object result = mapper.map(resultSet, uowMock);
+        assertNull(result);
     }
 }